requirements.txt
streamlit
openai-whisper
pydub
python-dotenv
google-generativeai
